notifier: extract notifier code to a separate file
