+++ server.go
@@ -7,7 +7,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"time"
 )
 
 // MaxEIRPacketLength is the maximum allowed AdvertisingPacket
@@ -422,54 +421,4 @@ func (c *conn) MTU() int           { return int(c.server.l2cap.mtu) }
 func (c *conn) UpdateRSSI() (rssi int, err error) {
 	// TODO
 	return 0, errors.New("not implemented yet")
-}
-
-type notifier struct {
-	l2c    *l2cap
-	char   *Characteristic
-	maxlen int
-	donemu sync.RWMutex
-	done   bool
-	// This throttle prevents multiple subsequent notifications from
-	// stepping on each others' toes. This toe-stepping appears to
-	// happen at both the HCI and the link layer.
-	throttle *time.Ticker
-}
-
-func newNotifier(l2c *l2cap, c *Characteristic, maxlen int) *notifier {
-	return &notifier{
-		l2c:      l2c,
-		char:     c,
-		maxlen:   maxlen,
-		throttle: time.NewTicker(50 * time.Millisecond),
-	}
-}
-
-func (n *notifier) Write(data []byte) (int, error) {
-	if n.Done() {
-		return 0, errors.New("central stopped notifications")
-	}
-	<-n.throttle.C
-	if err := n.l2c.sendNotification(n.char, data); err != nil {
-		return 0, err
-	}
-	return len(data), nil
-}
-
-func (n *notifier) Cap() int {
-	return n.maxlen
-}
-
-func (n *notifier) Done() bool {
-	n.donemu.RLock()
-	done := n.done
-	n.donemu.RUnlock()
-	return done
-}
-
-func (n *notifier) stop() {
-	n.donemu.Lock()
-	n.done = true
-	n.donemu.Unlock()
-	n.throttle.Stop()
 }
